chore: move config printing functions to separate file
