+++ pkg/printer/print.go
@@ -358,146 +358,6 @@ func Settings(settings models.WorkplaceSettings, jsonFlag bool) {
 	Table([]string{"id", "name", "billing email"}, rows, TableOptions())
 }
 
-// ScopedConfig print scoped config
-func ScopedConfig(conf models.ScopedOptions, jsonFlag bool) {
-	ScopedConfigSource(conf, jsonFlag, false)
-}
-
-// ScopedConfigSource print scoped config with source
-func ScopedConfigSource(conf models.ScopedOptions, jsonFlag bool, source bool) {
-	pairs := models.ScopedPairs(&conf)
-
-	if jsonFlag {
-		confMap := map[string]map[string]string{}
-
-		for name, pair := range pairs {
-			if *pair != (models.ScopedOption{}) {
-				scope := pair.Scope
-				value := pair.Value
-
-				if confMap[scope] == nil {
-					confMap[scope] = map[string]string{}
-				}
-				confMap[scope][name] = value
-			}
-		}
-
-		JSON(confMap)
-		return
-	}
-
-	var rows [][]string
-
-	for name, pair := range pairs {
-		if *pair != (models.ScopedOption{}) {
-			row := []string{name, pair.Value, pair.Scope}
-			if source {
-				row = append(row, pair.Source)
-			}
-			rows = append(rows, row)
-		}
-	}
-
-	// sort by name
-	sort.Slice(rows, func(a, b int) bool {
-		return rows[a][0] < rows[b][0]
-	})
-
-	headers := []string{"name", "value", "scope"}
-	if source {
-		headers = append(headers, "source")
-	}
-
-	Table(headers, rows, TableOptions())
-}
-
-// ScopedConfigValues print scoped config value(s)
-func ScopedConfigValues(conf models.ScopedOptions, args []string, values map[string]*models.ScopedOption, jsonFlag bool, plain bool, copy bool) {
-	if plain || copy {
-		vals := []string{}
-		for _, arg := range args {
-			if option, exists := values[arg]; exists {
-				vals = append(vals, option.Value)
-			}
-		}
-
-		print := strings.Join(vals, "\n")
-		if copy {
-			if err := utils.CopyToClipboard(print); err != nil {
-				utils.HandleError(err, "Unable to copy to clipboard")
-			}
-		}
-
-		if plain {
-			fmt.Println(print)
-			return
-		}
-	}
-
-	if jsonFlag {
-		filteredMap := map[string]string{}
-		for _, arg := range args {
-			if option, exists := values[arg]; exists {
-				filteredMap[arg] = option.Value
-			}
-		}
-
-		JSON(filteredMap)
-		return
-	}
-
-	var rows [][]string
-	for _, arg := range args {
-		if option, exists := values[arg]; exists {
-			rows = append(rows, []string{arg, option.Value, option.Scope})
-		}
-	}
-	Table([]string{"name", "value", "scope"}, rows, TableOptions())
-}
-
-// Configs print configs
-func Configs(configs map[string]models.FileScopedOptions, jsonFlag bool) {
-	if jsonFlag {
-		JSON(configs)
-		return
-	}
-
-	var rows [][]string
-	for scope, conf := range configs {
-		pairs := models.Pairs(conf)
-
-		for name, value := range pairs {
-			if value != "" {
-				rows = append(rows, []string{name, value, scope})
-			}
-		}
-	}
-
-	// sort by scope, then by name
-	sort.Slice(rows, func(a, b int) bool {
-		if rows[a][2] != rows[b][2] {
-			return rows[a][2] < rows[b][2]
-		}
-		return rows[a][0] < rows[b][0]
-	})
-
-	Table([]string{"name", "value", "scope"}, rows, TableOptions())
-}
-
-// ConfigOptionNames prints all supported config options
-func ConfigOptionNames(options []string, jsonFlag bool) {
-	if jsonFlag {
-		JSON(options)
-		return
-	}
-
-	rows := [][]string{}
-	for _, option := range options {
-		rows = append(rows, []string{option})
-	}
-	Table([]string{"name"}, rows, TableOptions())
-}
-
 // ConfigServiceTokensInfo print info of multiple config service tokens
 func ConfigServiceTokensInfo(tokens []models.ConfigServiceToken, number int, jsonFlag bool) {
 	maxTokens := int(math.Min(float64(len(tokens)), float64(number)))
